requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.21.1
SQLAlchemy==2.0.22
alembic==1.11.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.11
python-multipart==0.0.6
boto3==1.26.0
